Fix automatic template selection and add shortlist templates
The automatic template selection has been changed to use the `.main`
special variable, so the various libraries (e.g. `facebook`, `video`,
etc.) need to be in the main scope, imported in `simple.ftl`.

Also added per result type shortlist template, based on the collection
the shortlisted item is coming from. This is only possible by looking at
the `C` metadata since the `item.collection` property always point to
the meta collection (as it has been added in the shortlist from the meta
collection...)