Remove scripts and data to generate fake analytics
We can't keep them in because this collection is cloned to deploy HE
packages for clients, and it will be used by partners in this way to.

Additionally the scripts are already in `hosted_bin/` on SaaS so there's
no need to check them in the collection.

Updated `.gitignore` to ignore the query files so that we can put them
in on the Showcase server without having them on the collection files
and have them cloned when deploying HE packages.