Attempt to fool proof the direct use of Stencil resources
Files from the Stencils should be copied in when deploying the HE
package, otherwise changes in the Stencils will likely break collections
that include the resources.

This is easy to overlook though, so to try to make it fool proof,
perform a check and display a warning message if resources are directly
included and the collection is not the showcase one (the only one where
this is permitted).