Add a border to tabs
We often end up having to add a border because the background color of
the search form is usually much lighter than the showcase (i.e. white or
gray instead of dark red).

Adding the border properly is always fiddly, so just add a default
border on the showcase collection that can be reused.

Used a neutral, translucent border with the hope that it will
automatically work in most situations. If not, adjusting the color
should be enough to tweak/fix it.