[FUNNELAPPS-540] Remove macro calls and cleanup templates
Funnelback 15.20 added a listMetadata property on the result object that can now be used to access metadata items in a cleaner way than the raw string. The macros previously used for that are no longer necessary, and using the standard Freemarker syntax for sequences makes the templates more readable.

Also uses trailing commas in the Javascript section to make it a bit more readable.