[FUNNELAPPS-413] Rename Showcase service image
Changing the name of the image for the service which shows in the Marketing Dashboard is a step that is very often forgotten by implementers in demos and projects. If the name is different, it is more likely people will notice it. Ultimately, this allows for an automated solution to check implementations before they go live (perhaps using CI/CD)