[FUNNELAPPS-456] Add a template for embedded HTML integration
Adds a template for embedded HTML integration that does not include elements such as <html>, <head>, <body>, etc. The work to update simple.ftl for this integration method is repetitive and manual work on every project. This provides a new template file that will be automatically copied over simple.ftl in the setup process for these projects.